# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NIDS Sentinel API",
    "description": "AI-Powered Real-Time Network Intrusion Detection and Threat Analysis API.\n\nProvides endpoints for authentication, real-time traffic ingestion, threat alert management, GeoIP lookup, ML model metrics, and PCAP file analysis.\n\n**WebSocket endpoint** (not representable in OAS 3.0): `ws://localhost:8000/api/v1/traffic/stream?token=<jwt>` — streams live classified packet data to connected dashboard clients. Connection is closed with code 4001 if token is missing or 4003 if invalid.",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
...
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/v1/alerts/recent": {
      "get": {
        "operationId": "getRecentAlerts",
        "summary": "Fetch recent threat alerts",
        "tags": ["Alerts"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50, "maximum": 200 },
            "description": "Maximum number of alerts to return (capped at 200)"
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Filter results by threat label (e.g. DoS, Probe, U2R (Root Access))"
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered list of recent threat log entries (newest first)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ThreatLogEntry" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized — missing or invalid credentials" },
          "403": { "description": "Forbidden — could not validate credentials" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/v1/alerts/stats": {
      "get": {
        "operationId": "getAlertStats",
        "summary": "Get aggregated threat statistics",
        "description": "Returns total threat count, counts grouped by label, and the top 10 attacker source IPs.",
        "tags": ["Alerts"],
        "responses": {
          "200": {
            "description": "Aggregated threat statistics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AlertStats" }
              }
            }
          },
          "401": { "description": "Unauthorized — missing or invalid credentials" },
          "403": { "description": "Forbidden — could not validate credentials" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/v1/alerts/timeline": {
      "get": {
        "operationId": "getThreatTimeline",
        "summary": "Get threat counts bucketed by time period for the historical chart",
        "tags": ["Alerts"],
        "parameters": [
          {
            "name": "time_range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["24h", "7d", "30d"],
              "default": "24h"
            },
            "description": "Time window: 24h = hourly buckets, 7d = daily buckets, 30d = daily buckets"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of time-bucketed threat and traffic counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TimelineEntry" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized — missing or invalid credentials" },
          "403": { "description": "Forbidden — could not validate credentials" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/v1/alerts/explain/{label}": {
      "get": {
        "operationId": "explainThreat",
        "summary": "AI explainability — why the model flagged a given threat type",
        "description": "Returns the key contributing features, severity rating, and recommended mitigation for a threat label. Data is derived from Random Forest feature importances.",
        "tags": ["Alerts"],
        "parameters": [
          {
            "name": "label",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["DoS", "DDoS (Ping of Death)", "Probe", "U2R (Root Access)"]
            },
            "description": "Threat label to explain"
          }
        ],
        "responses": {
          "200": {
            "description": "Threat explanation with key indicators, severity, and mitigation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ThreatExplanation" }
              }
            }
          },
          "401": { "description": "Unauthorized — missing or invalid credentials" },
          "403": { "description": "Forbidden — could not validate credentials" },
          "404": { "description": "No explanation available for the given label" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/v1/alerts/export": {
      "get": {
        "operationId": "exportAlertsPdf",
        "summary": "Export the most recent 100 threat alerts as a downloadable PDF report",
        "tags": ["Alerts"],
        "responses": {
          "200": {
            "description": "PDF threat report",
            "headers": {
              "Content-Disposition": {
                "schema": { "type": "string" },
                "description": "attachment; filename=nids-threat-report.pdf"
              }
            },
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": { "description": "Unauthorized — missing or invalid credentials" },
          "403": { "description": "Forbidden — could not validate credentials" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/v1/alerts/geoip/{ip}": {
      "get": {
        "operationId": "geoipLookup",
        "summary": "Server-side GeoIP lookup for an IP address",
        "description": "Proxies the ip-api.com API to avoid CORS restrictions. Results are cached in-process (up to 1000 IPs). Returns the raw ip-api.com JSON response.",
        "tags": ["Alerts"],
        "parameters": [
          {
            "name": "ip",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "IPv4 or IPv6 address to look up"
          }
        ],
        "responses": {
          "200": {
            "description": "GeoIP data from ip-api.com (fields include status, country, countryCode, region, city, lat, lon, isp, query, etc.)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GeoIPResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized — missing or invalid credentials" },
          "403": { "description": "Forbidden — could not validate credentials" },
          "502": { "description": "Bad gateway — upstream GeoIP lookup failed or timed out" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/block": {
      "post": {
        "operationId": "blockThreat",
        "summary": "Mark a threat as blocked and ban the source IP on the OS firewall",
        "description": "Sets `is_blocked = true` on the ThreatLog entry and invokes the firewall service (pfctl/iptables) to block the source IP.",
        "tags": ["Alerts"],
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID of the alert to block"
          }
        ],
        "responses": {
          "200": {
            "description": "Threat marked blocked; firewall status indicates whether the OS-level ban succeeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BlockResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized — missing or invalid credentials" },
          "403": { "description": "Forbidden — could not validate credentials" },
          "404": { "description": "Alert not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256-signed JWT obtained from POST /api/v1/login/access-token. Token expires after 90 days."
      }
    },
    "schemas": {
...
      "ThreatLogEntry": {
        "type": "object",
        "required": ["id", "src_ip", "dst_ip", "protocol", "label", "confidence", "is_blocked"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "timestamp": { "type": "string", "format": "date-time", "nullable": true },
          "src_ip": { "type": "string" },
          "dst_ip": { "type": "string" },
          "protocol": { "type": "string" },
          "label": { "type": "string" },
          "confidence": { "type": "number", "format": "float" },
          "is_blocked": { "type": "boolean" }
        }
      },
      "AlertStats": {
        "type": "object",
        "required": ["total_threats", "by_label", "top_sources"],
        "properties": {
          "total_threats": { "type": "integer" },
          "by_label": {
            "type": "object",
            "additionalProperties": { "type": "integer" },
            "description": "Threat counts keyed by label name"
          },
          "top_sources": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TopSource" }
          }
        }
      },
      "TopSource": {
        "type": "object",
        "required": ["ip", "count"],
        "properties": {
          "ip": { "type": "string" },
          "count": { "type": "integer" }
        }
      },
      "TimelineEntry": {
        "type": "object",
        "required": ["time", "threats", "traffic"],
        "properties": {
          "time": {
            "type": "string",
            "description": "Human-readable time label (e.g. '14:00' for 24h, 'Mon' for 7d, 'Jun 01' for 30d)"
          },
          "threats": { "type": "integer" },
          "traffic": { "type": "integer" }
        }
      },
      "ThreatKeyIndicator": {
        "type": "object",
        "required": ["feature", "impact", "detail"],
        "properties": {
          "feature": { "type": "string" },
          "impact": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "CRITICAL"]
          },
          "detail": { "type": "string" }
        }
      },
      "ThreatExplanation": {
        "type": "object",
        "required": ["label", "description", "key_indicators", "severity", "mitigation"],
        "properties": {
          "label": { "type": "string" },
          "description": { "type": "string" },
          "key_indicators": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ThreatKeyIndicator" }
          },
          "severity": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "CRITICAL"]
          },
          "mitigation": { "type": "string" }
        }
      },
      "BlockResponse": {
        "type": "object",
        "required": ["status", "id", "ip", "firewall_active"],
        "properties": {
          "status": { "type": "string", "example": "blocked" },
          "id": { "type": "string", "format": "uuid" },
          "ip": { "type": "string" },
          "firewall_active": { "type": "boolean" }
        }
      },
      "GeoIPResponse": {
        "type": "object",
        "additionalProperties": true,
        "description": "Raw response from ip-api.com. Common fields: status (string), country (string), countryCode (string), region (string), regionName (string), city (string), lat (number), lon (number), isp (string), query (string)."
      },
...
    }
  }
}